feat: Add timeout to toast display function
