+++ calc/toolcommands.js
@@ -1,4 +1,4 @@
-import { $ } from "../src/utilities.js";
+import { $, displayToast } from "../src/utilities.js";
 import { STAT_LISTS, COUNT_PREFIX, CLIP_TEXT } from "../src/constants.js";
 
 $("#btn-clipboard").addEventListener("click", copyResults);
@@ -19,31 +19,12 @@ function copyResults() {
 
   navigator.clipboard.writeText(text.join("\n")).then(
     () => {
-      displayToast(true);
+      displayToast("복사되었습니다.", 1200);
     },
     () => {
-      displayToast(false);
+      displayToast("복사에 실패했습니다.", 1200);
     }
   );
-}
-
-function displayToast(isSuccessful) {
-  const container = $("#main-content");
-  if (container.lastElementChild.id === "toast") {
-    container.lastElementChild.remove();
-  }
-
-  const newToast = document.createElement("div");
-  newToast.id = "toast";
-  newToast.innerText = isSuccessful
-    ? "복사되었습니다."
-    : "복사에 실패했습니다.";
-
-  container.appendChild(newToast);
-  newToast.style.animationPlayState = "running";
-  setTimeout(() => {
-    newToast.remove();
-  }, 610);
 }
 
 function addFavorites() {
